changed divide to double

+++ src/main/java/com/nbs/calculator/Calculator.java
@@ -15,7 +15,7 @@ public class Calculator {
 
 	}
 
-	public int divide(int number1, int number2) {
+	public double divide(double number1, double number2) {
 		return number1 / number2;
 	}
